Added the __build__.py to merges the individual source files.
The target file (in which the final code will be written, and hence
needs to be specified) can now be shared with others easily.
Also, the --loop option watches for changes to perform auto-updates.

Since the merge process involves simple concatenation of files:
(1) All required external modules need to be specified here.
(2) As long as the concatenation order is properly specified,
    modules need not import one another.

PS: Isn't this awesome? :D