__build__.py implements node.js style 'exports' dictionary.
In the process of wrapping each file's contents with a function (to create a new
scope), each line is being prefixed with a horizontal tab character. Since this
process can corrupt multi-line string, they must not be used.

Additionally, since multi-line strings are the only place where blank lines matter,
given that the former will not be used anyway, all blank lines will be discarded.