__build__.py can now debounce events.
As to many events were being fired while updating files, the debouncing
procedure ensures a quiet period before updating the target file. This
reduces unnecessary updates of the latter.

Apparently, even if you don't make changes in a file, but just save it as
it is, it counts as a modification (as the last-modified-time gets updated).