Added the TaskGroup class with prettytable tabulation.
Additionally, the following updates were made:
(1) Added a number of new static methods to the Date class.
(2) Disabled direct access to properties of Task instances and added support for operators.
(3) Renamed scan() to load() in the TaskFile class, and added group() and select() methods.

Sorry ... I know it's too much. But while working, the only thing on my mind was to
implement the TaskFile.select() method. Added whatever was needed for that purpose,
occasionally pausing to fix bugs and abstraction leaks.