Added the birthday tag & the birthdays group.
Had to add the Task.__nostatus() function that returns the raw task text,
but without status tags, and including (and replacing) the ones specified
in the argument. It is used in Task.periodic & Task.birthday methods.

Also had to update TaskFile.__process() to handle birthday tags.

This is an attempt to make me feel guilty about not calling friends and
family on their birthdays.